% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.model.R
\name{check.model}
\alias{check.model}
\title{A visualization tool of diagnostics of fitted model in the partial association analysis}
\usage{
check.model(object, output = c("qq", "fitted", "covariate"), ...)
}
\arguments{
\item{object}{The object in "PAsso" class that is generated by "test" or "PAsso".}

\item{output}{A character string specifying what output to plot. Default is \code{"qq"} which
produces a quantile-quantile plots of the residuals. \code{"fitted"} will give a matrix
plot between residuals and fitted response. \code{"covariates"} will give a matrix plot
between each residual and each covariate.}

\item{...}{Additional optional arguments for drawing plots to be passed onto.}
}
\value{
A list contains plots of \code{"ggplot"} object
}
\description{
A plot matrix to display the QQ-plot of each fitted model when conducting the partial
association analyses. Based on the input of "PAsso" object, each plot stands for the
QQ-plot of the surrogate residuals of fitted models. All the plots are based on the
surrogate residuals generated from \code{"resides"} function.Graphics are designed
based on \code{\link{ggplot2}}.
}
\examples{
# import data -------------------------------------------------------------
data("nes2016_pre")
# "PAsso" function: input three responses ----------------------------
PAsso_2 <- PAsso(responses = c("Prevote.num", "PID", "selfLR"),
                adjustments = c("income.num", "age", "edu.year"),
                data = nes2016)
summary(PAsso_2, digits=4)

check_qq <- check.model(object = PAsso_2, color="blue",output = "qq")

check_fitted <- check.model(object = PAsso_2, output = "fitted")

check_covar <- check.model(object = PAsso_2, output = "covariate")

}
