% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Correlation coefficients and partial correlation coefficients between
ordinal responses after adjusting for a set of covariates.}
\usage{
corr(responses, adjustments, data, association = c("partial",
  "marginal"), models = c("probit", "probit"), method = c("kendall",
  "pearson", "wolfsigma"), resids.method = c("latent", "jitter", "Sign",
  "General", "Deviance"), fitted.models = NULL, rep_num = 100, ...)
}
\arguments{
\item{responses}{A string vector that specifies response variables. It requires to be equal
or greater than two variables in the data frame.}

\item{adjustments}{A string vector specifies covariates/confounders that need to
be adjusted.}

\item{data}{the data.frame including responses and adjustments.}

\item{association}{A character string indicating whether the partial association
coefficient \code{"partial"} (default) or the marginal correlation \code{"marginal"}
is to be computed. They can be abbreviated.}

\item{models}{A string vector contains default link function of fitting models with
respect to each response variable.}

\item{method}{A string argument to specify correlation coefficient method.
Three choices \code{c("kendall", "pearson", "wolfsigma")}. The default is
\code{"kendall"}}

\item{resids.method}{A character string specifying which method to use to generate the
residuals. Current options are \code{"latent"} and
\code{"jitter"}. Default is \code{"latent"}.
\describe{
  \item{\code{latent}}{surrogate approach;}
  \item{\code{jitter}}{jittering approach;}
  \item{\code{Sign}}{Sign-based residuals;}
  \item{\code{General}}{Generalized residuals (Franses and Paap 2001);}
  \item{\code{Deviance}}{Deviance residuals (-2*loglik).}
}}

\item{fitted.models}{A list contains all the models (S3 objects) you want to
assess for the partial association between ordinal responses after adjusting
for a set of covariates covariates. All of these models should be applied to the
same dataset, having same covariates, same sample size etc. The models in this
list can be an object of class \code{\link[ordinal]{clm}},
\code{\link[stats]{glm}}, \code{\link[rms]{lrm}}, \code{\link[rms]{orm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{rep_num}{A number to specify repeat times of simulation of surrogate residuls
such that the partial correlation coefficients are calculated repeatly. The final
correlation coefficents the average of all partial correlation coefficients.
It is the \code{"nsim"} argument in \code{"resids()"} function.}

\item{...}{Additional optional arguments. (say \code{"method"} and \code{"jitter.scale"}
in \code{"resids"} function. Just keep them as default if not specified. Currently ignored.)}
}
\value{
An object of class \code{"PAsso"} is a list containing at least the following
components if \code{"partial"} is the association argument. It contains the partial
correlation matrix and multiple repeats if \code{rep_num} > 1. This object has "arguments"
attributes saved as c(association, method, resids.method).
The list contains:
\describe{
  \item{\code{corr}}{The estimated correlation matrix(average of \code{rep_MatCorr}) of partial association after
  adjusting confounders;}
  \item{\code{rep_corr}}{The replications of estimated correlation matrix;}
  \item{\code{rep_SRs}}{The replications of surrogate residuals if partial association is applied;}
  \item{\code{fitted.models}}{The list stores all fitted.models;}
  \item{\code{data}}{The data.frame of original dataset;}
  \item{\code{mods_n}}{The sample size of each fitted model;}
  \item{\code{cor_func}}{The correlation function after assign different method.}
}
}
\description{
This function is mainly designed to conduct partial association analysis
partial association analysis bewteen two or more ordinal response variables
after adjusting for a set of covariates. It can conduct marginal correlation
analyses without adjusting covariates. It generates an object that has
correlation coefficients matrix, and some attributes ("arguments" saves
c(association, method, resids.method)). The attribute "responses" contains
the response variables.
}
\examples{
# import data
data(nes96)
PAsso_1 <- corr(responses = c("vote.num", "PID"),
adjustments = c("income.num", "age", "edu.year"),
data = nes96,
association = c("partial"),
models = c("probit", "probit"),
method = c("kendall"),
resids.method = "latent", fitted.models = NULL, rep_num = 100)

# Compare marginal correlation and partial correlation.
#tau; tau.sd.boot
PAsso_1$corr;
#Pcor_2$corr;


}
