% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.test.R
\name{corr.test}
\alias{corr.test}
\title{corr.test}
\usage{
corr.test(object, boot_SE = 300, H0 = 0, parallel = FALSE)
}
\arguments{
\item{object}{an object of "PartialCor" class, which is generated by "Pcor_SR.def" function.}

\item{boot_SE}{bootstrap replications. It may be very slow for partial association
so a double-check question should be answered before move on.}

\item{H0}{null hypothesis of partial correlation coefficient.}

\item{parallel}{logical argument whether conduct parallel for bootstrapping partial association.}
}
\description{
corr.test
}
