% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_AdjCat}
\alias{df_AdjCat}
\title{A dataset with ordinal response generated from an adjacent categories model}
\format{
A data frame with 10000 rows and 11 variables.
\itemize{
  \item \code{X} A continuous predictor variable.
  \item \code{Y1} The response variable; an ordered factor.
  \item \code{Y2} The response variable; an ordered factor.
}
}
\usage{
data(df_AdjCat)
}
\description{
Data simulated from an adjacent categories regression model
with an ordered (preferably) factor response. beta1 = 1, beta2 = -1,
alpha1 = (-Inf, -3, -2, 0, 2, 3, Inf), alpha2 = (-Inf, -2, 0, 2, Inf)
}
\examples{
#
# Adjacent Categories Regression Model Example to compare different residuals
#

data("df_AdjCat")
summary(df_AdjCat$data)
fit_clm1 <- VGAM::vglm(Y1 ~ X, family =
                      VGAM::cumulative(link = "logit",reverse=TRUE,parallel = TRUE),
                      data = df_AdjCat$data)
fit_clm2 <- VGAM::vglm(Y2 ~ X, family =
                       VGAM::cumulative(link = "logit",reverse=TRUE,parallel = TRUE),
                       data = df_AdjCat$data)
SR1 <- resids(fit_clm1, method = "latent",boot_id = NULL)
SR2 <- resids(fit_clm2, method = "latent", boot_id = NULL)

## obtain SBC residuals (Li and Shepherd 2012 JASA/Biometrika)
PR1 <- resids(fit_clm1, method = "Sign", boot_id = NULL)
PR2 <- resids(fit_clm2, method = "Sign", boot_id = NULL)

## obtain generalized residuals (Franses and Paap 2001 book)
GR1 <- resids(fit_clm1, method = "General", boot_id = NULL)
GR2 <- resids(fit_clm2, method = "General", boot_id = NULL)

## obtain deviance residuals
DR1 <- resids(fit_clm1, method = "Deviance", boot_id = NULL)
DR2 <- resids(fit_clm2, method = "Deviance", boot_id = NULL)

## visualize residual vs. residual
par(mfrow=c(2,2))
par(mar=c(4, 4.8, 2.5, 1.5))

plot(PR1, PR2, pch=".", main = "Sign-based Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")))
plot(GR1, GR2, pch=".", main = "Generalized Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")), xlim = c(-4,4), ylim=c(-4,4))
plot(DR1, DR2, pch='.', main = "Deviance Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")))
plot(SR1, SR2, pch=".", main = "Surrogate Residuals", xaxt="n", yaxt="n",
     xlab = expression(R[1]), ylab = expression(R[2]),
     xlim = c(-1/2,1/2), ylim=c(-1/2,1/2))
axis(1, at=seq(-0.5, 0.5, 0.25), labels = seq(-0.5, 0.5, 0.25))
axis(2, at=seq(-0.5, 0.5, 0.25), labels = seq(-0.5, 0.5, 0.25))

}
\references{
Liu, Dungang, Li, Shaobo, Yu, Yu, and Moustaki, Irini. Assessing partial association between
ordinal variables: quantification, visualization, and hypothesis testing, \emph{Journal of the
American Statistical Association}, Revision under review.
}
\keyword{datasets}
