% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ParA}
\alias{df_ParA}
\title{Ordinal Response with Adjacent Categories Probabilities}
\format{A data frame with 10000 rows and 11 variables.
\itemize{
  \item \code{X} A continuous predictor variable.
  \item \code{Y1} The response variable; an ordered factor.
  \item \code{Y2} The response variable; an ordered factor.
}}
\usage{
data(df_ParA)
}
\description{
Data simulated from an adjacent categories regression model
with an ordered (preferably) factor response. beta1 = 1, beta2 = -1,
alpha1 = (-Inf, -3, -2, 0, 2, 3, Inf), alpha2 = (-Inf, -2, 0, 2, Inf)
}
\examples{
head(df_ParA)
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association}.
}
\keyword{datasets}
