% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{generate_residuals_acat}
\alias{generate_residuals_acat}
\title{generate_residuals_acat}
\usage{
generate_residuals_acat(y, X, alpha, beta, nsim = 1)
}
\arguments{
\item{y}{A vector inputs the response variable.}

\item{X}{A data.frame inputs the covariates.}

\item{alpha}{A vector provides the estimated intercepts of adjacent categories model. If the response
has k levels, there should be k+1 numbers in this alpha argument with the k-1 estimated intercepts.
The lower bound and upper bound are "-Inf" and "Inf".}

\item{beta}{A vector provides the estimated coefficients.}

\item{nsim}{A number to specify the replication of residuals.}
}
\value{
A vector or a matrix (nsim>1) of residuals for the adjacent categories model.
}
\description{
generate_residuals_acat
}
\keyword{internal}
