% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairs.PAsso}
\alias{ggpairs.PAsso}
\title{A visualization tool of the partial association analyses (plot matrix)}
\usage{
ggpairs.PAsso(object, color = "blue", ...)
}
\arguments{
\item{object}{the object in "PAsso" class that is generated by "corr.test" or "corr".}

\item{color}{spedify the points color.}

\item{...}{Additional optional arguments to be passed onto.}
}
\value{
A \code{"GGally"} object.
}
\description{
A plot matrix to display the results of partial association analyses. Upper-triangle contains
scatter-plot matrix between each pair of response variables. Lower-triangle contains the
partial correlation coefficients adjusted by covariates.
}
\details{
A pairwise plot matrix reveals the partial association between ordinal variables.
All the plots are based on surrogate residuals generated from \code{"resides"} function.
Graphics are designed based on \code{\link{ggplot2}} and \code{"GGally"}.
}
\examples{
# See ?resids for an example
?resids
}
