% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairs.resid}
\alias{ggpairs.resid}
\title{Partial regression plot matrix}
\usage{
ggpairs.resid(resid_Mat, color = "blue", ...)
}
\arguments{
\item{resid_Mat}{An matrix contains residuals for all ordinal responses.}

\item{color}{spedify the points color}

\item{...}{Additional optional arguments to be passed onto
\code{\link[sure]{resids}}.}
}
\value{
A \code{"GGally"} object.
}
\description{
A pairwise plot matrix reveals the partial association between ordinal variables.
All the plots are based on surrogate residuals generated from \code{"resides"} function.
Graphics are designed based on \code{\link[ggplot2]{ggplot2}} and
\code{\link[GGally]{GGally}}.
}
\examples{
# See ?resids for an example
?resids
}
