% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.arrange.R
\name{grid.arrange}
\alias{grid.arrange}
\title{Arrange multiple grobs on a page}
\usage{
grid.arrange(..., newpage = TRUE)
}
\description{
See \link[gridExtra]{arrangeGrob} for more details.
}
\keyword{internal}
