% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nes2016.clean.R
\name{nes2016.clean}
\alias{nes2016.clean}
\title{nes2016pre.clean
is a function to clean the nes2016raw and return the cleaned dataset
with the pre-election interview only.}
\usage{
nes2016.clean(interview = c("pre", "post", "both"))
}
\arguments{
\item{interview}{A string to specify the obsertations that want to be include. Default is "pre",
which only keep the observations without missing in the pre-election interview. "post" will return
observations without missing in the post-election interview. If interview equals to "both",
returned dataset only contains the observations without missing in both the pre and post-election
interviews.}
}
\value{
a dataset with pre-election intend vote and demographic information.
}
\description{
nes2016pre.clean
is a function to clean the nes2016raw and return the cleaned dataset
with the pre-election interview only.
}
\examples{
nes2016_pre <- nes2016pre.clean()

}
