% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nes2016_prepost}
\alias{nes2016_prepost}
\title{US 2016 national election study of respondents without missing in both pre-election and post-election.}
\format{
A data frame with 1690 rows and 14 variables.
\itemize{
  \item \code{age} Respondent's age in years.

  \item \code{edu.year} Respondent's education year, which is mapped from raw data
  \code{nes2016_raw}. \code{MS}=8, \code{HSdrop}=11, \code{HS}=12, \code{Coll}=14,
  \code{CCdeg}=15, \code{BAdeg}=17, \code{MAdeg}=19.

  \item \code{income.num} Respondent's family income in thousands: an numerical variable.

  \item \code{PID} Party identification: a numeric variable with value from 1 to 7
  representing strong Democrat, \code{strDem} < weak Democrat, \code{weakDem} <
  independent Democrat, \code{indDem} < independent independent \code{indind} <
  indepedent Republican, \code{indRep} < weak Republican, \code{weakRep} <
  strong Republican, \code{strRep}.
  \item \code{selfLR} Left-Right self-placement of respondent: a numeric variable
  with value from 1 to 7 representing \code{extLib}: extremely liberal, \code{Lib}: liberal,
  \code{sliLib}: slightly liberal, \code{Mod}: moderate, \code{sliCon}:
  slightly conservative, \code{Con}: conservative, \code{extCon}: extremely
  conservative. \code{extLib} < \code{Lib} < \code{sliLib} <
  \code{Mod} < \code{sliCon} < \code{Con} < \code{extCon}.

  \item \code{TrumpLR} Left-Right placement of Donald Trump (same
  scale as selfLR), a numeric variable with value from 1 to 7.

  \item \code{ClinLR} Left-Right placement of Bill Clinton (same
  scale as selfLR): a numeric variable with value from 1 to 7.

  \item \code{IntendVote} The intend vote two months preceeding the
  November election (Pre-election interview). It is a factor with levels
  \code{HillaryClinton} and \code{DonaldTrump}.

  \item \code{Prevote.num} Recode the intend vote to numeric values,
  'HillaryClinton'=0, 'DonaldTrump'=1.

  \item \code{voteResult} The actual vote answered in the interview during the
  two months following the election (Post-election interview). It is a factor
  with levels \code{HillaryClinton} and \code{DonaldTrump}.

  \item \code{Postvote.num} Recode the actual vote to numeric values,
  'HillaryClinton'=0, 'DonaldTrump'=1.

  \item \code{WeightforPreElection} Pre-election weight of a respondent.

  \item \code{WeightforPostElection} Post-election weight of a respondent.

}
}
\usage{
data(nes2016_prepost)
}
\description{
A subset of the 2016 American National Election Study. Pre-election preference is recorded
as "IntendVote", while the actual vote is "voteResult".
"Prevote.num" and "Postvote.num" are the numeric for them, respectively. Observations
with missing values, or "No thought" responses have been removed. Responsdents expressing
a voting preference other than Clinton or Trump have been removed.
}
\examples{
head(nes2016_prepost)
}
\keyword{datasets}
