% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nes2016_raw}
\alias{nes2016_raw}
\title{Raw dataset of US 2016 national election study}
\format{
A data frame with 4271 rows and 12 variables.
\itemize{
  \item \code{WeightforPreElection} Pre-election weight of a respondent.

  \item \code{WeightforPostElection} Post-election weight of a respondent.

  \item \code{selfLR} Left-Right self-placement of respondent: an ordered
  factor with levels \code{extLib}: extremely liberal, \code{Lib}: liberal,
  \code{sliLib}: slightly liberal, \code{Mod}: moderate, \code{sliCon}:
  slightly conservative, \code{Con}: conservative, \code{extCon}: extremely
  conservative. \code{extLib} < \code{Lib} < \code{sliLib} <
  \code{Mod} < \code{sliCon} < \code{Con} < \code{extCon}.

  \item \code{ClinLR} Left-Right placement of Bill Clinton (same
  scale as selfLR): an ordered factor with levels \code{extLib} <
  \code{Lib} < \code{sliLib} < \code{Mod} < \code{sliCon} <
  \code{Con} < \code{extCon}.

  \item \code{PID} Party identification: an ordered factor with levels strong
  Democrat, \code{strDem} < weak Democrat, \code{weakDem} < independent Democrat,
  \code{indDem} < independent independent \code{indind} < indepedent Republican,
  \code{indRep} < weak Republican, \code{weakRep} < strong Republican, \code{strRep}.

  \item \code{agegroup} Respondent's age category.

  \item \code{age} Respondent's age in years.

  \item \code{educ} Respondent's education category: an ordered factor with levels
  8 years or less, \code{MS} < high school dropout, \code{HSdrop} < high school
  diploma or GED, \code{HS} < some College, \code{Coll} < Community or junior
  College degree, \code{CCdeg} < BA degree, \code{BAdeg} < postgraduate degree,
  \code{MAdeg}.

  \item \code{income} Categorical variable representing the Respondent's family income

  \item \code{income.num} Respondent's family income in thousands: an numerical variable.
  income.num = recode(income, '(01) 01. Under $5,000' = 5, '(02) 02. $5,000-$9,999' = 7.5,
  '(03) 03. $10,000-$12,499' = 11.25, '(04) 04. $12,500-$14,999' = 13.75,
  '(05) 05. $15,000-$17,499' = 16.25, '(06) 06. $17,500-$19,999' = 18.75,
  '(07) 07. $20,000-$22,499' = 21.25, '(08) 08. $22,500-$24,999' = 23.75,
  '(09) 09. $25,000-$27,499' = 26.25, '(10) 10. $27,500-$29,999' = 28.75,
  '(11) 11. $30,000-$34,999' = 32.5, '(12) 12. $35,000-$39,999' = 37.5,
  '(13) 13. $40,000-$44,999' = 42.5, '(14) 14. $45,000-$49,999' = 47.5,
  '(15) 15. $50,000-$54,999' = 52.5, '(16) 16. $55,000-$59,999' = 57.5,
  '(17) 17. $60,000-$64,999' = 62.5, '(18) 18. $65,000-$69,999' = 67.5,
  '(19) 19. $70,000-$74,999' = 72.5, '(20) 20. $75,000-$79,999' = 77.5,
  '(21) 21. $80,000-$89,999' = 85, '(22) 22. $90,000-$99,999' = 95,
  '(23) 23. $100,000-$109,999' = 105, '(24) 24. $110,000-$124,999' = 117.5,
  '(25) 25. $125,000-$149,999' = 137.5, '(26) 26. $150,000-$174,999' = 162.5,
  '(27) 27. $175,000-$249,999' = 212.5, '(28) 28. $250,000 or more' = 250)

  \item \code{IntendVote} The intend vote two months preceeding the
  November election (Pre-election interview). It is a factor with levels
  \code{HillaryClinton} and \code{DonaldTrump}.

  \item \code{Prevote.num} Recode the intend vote to numeric values,
  'HillaryClinton'=0, 'DonaldTrump'=1.

  \item \code{voteResult} The actual vote answered in the interview during the
  two months following the election (Post-election interview). It is a factor
  with levels \code{HillaryClinton} and \code{DonaldTrump}.

  \item \code{Postvote.num} Recode the actual vote to numeric values,
  'HillaryClinton'=0, 'DonaldTrump'=1.
}
}
\usage{
data(nes2016_raw)
}
\description{
A subset of the 2016 American National Election Study.
Pre-election preference is recorded as "IntendVote", while the actual vote is "voteResult".
}
\examples{
head(nes2016_raw)
}
\keyword{datasets}
