% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasol.R
\docType{package}
\name{parasol}
\alias{parasol}
\alias{parasol-package}
\title{parasol: An R package for assessing the Partial ASSociation between
Ordinal variables. It includes quantification, visualization, and hypothesis testing.}
\description{
The \code{parasol}: a package for assessing the PARtial ASsociation between
OrdinaL variables for fitted ordinal and general (e.g., binary)
regression models of class
\code{\link[ordinal]{clm}}, \code{\link[stats]{glm}},
\code{\link[rms]{lrm}}, \code{\link[rms]{orm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.
}
\details{
The development version can be found on GitHub:
\url{https://github.com/XiaoruiZhu/parasol}. As of right now, \code{parasol} exports the
following functions:
\itemize{
  \item{\code{Pcor_SR2}} - calculate partial association based on
  surrogate residuals (or others in the future);
  \item{\code{resids}} - construct (surrogate-based) residuals;
  \item{\code{surrogate}} - construct sorrogate latent variables;
  \item{\code{autoplot}} - plot diagnostics using
  \code{\link[ggplot2]{ggplot2}}-based graphics;
  \item{\code{ggpairs.resid}} - draw partial regression plot matrix;
  \item{\code{gof}} - simulate p-values from a goodness-of-fit test.
}
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
