% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAsso.R
\name{PAsso}
\alias{PAsso}
\title{Partial association analysis between ordinal responses after adjusting for a set of covariates}
\usage{
PAsso(
  responses,
  adjustments,
  data,
  association = c("partial", "marginal"),
  uni.model = c("probit", "logit"),
  models = NULL,
  method = c("kendall", "pearson", "wolfsigma"),
  resids.method = c("latent", "jitter", "Sign", "General", "Deviance"),
  fitted.models = NULL,
  rep_num = 30,
  ...
)
}
\arguments{
\item{responses}{A string vector that specifies response variables. It requires to be equal
or greater than two variables in the data frame.}

\item{adjustments}{A string vector specifies covariates/confounders that need to
be adjusted.}

\item{data}{the data.frame including responses and adjustments.}

\item{association}{A character string indicating whether the partial association
coefficient \code{"partial"} (default) or the marginal correlation \code{"marginal"}
is to be computed. They can be abbreviated.}

\item{uni.model}{A character string specifying the universal model setting for all
responses. Default \code{"probit"} refers to cumulative probit model. \code{"logit"}}

\item{models}{A string vector contains default link functions of fitting models with
respect to each response variable. If \code{"uni.model"} is provided, this one will
be generated automaticlly.}

\item{method}{A string argument to specify correlation coefficient method.
Three choices \code{c("kendall", "pearson", "wolfsigma")}. The default is
\code{"kendall"}}

\item{resids.method}{A character string specifying which method to use to generate the
residuals. Current options are \code{"latent"} and
\code{"jitter"}. Default is \code{"latent"}.
\describe{
  \item{\code{latent}}{surrogate approach;}
  \item{\code{jitter}}{jittering approach;}
  \item{\code{Sign}}{Sign-based residuals;}
  \item{\code{General}}{Generalized residuals (Franses and Paap 2001);}
  \item{\code{Deviance}}{Deviance residuals (-2*loglik).}
}}

\item{fitted.models}{A list contains all the models (S3 objects) you want to
assess for the partial association between ordinal responses after adjusting
for a set of covariates covariates. All of these models should be applied to the
same dataset, having same covariates, same sample size etc. The models in this
list can be an object of class \code{\link[ordinal]{clm}},
\code{\link[stats]{glm}}, \code{\link[rms]{lrm}}, \code{\link[rms]{orm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{rep_num}{A number to specify repeat times of simulation of surrogate residuls
such that the partial correlation coefficients are calculated repeatly. The final
correlation coefficents the average of all partial correlation coefficients.
It is the \code{"nsim"} argument in \code{"resids()"} function.}

\item{...}{Additional optional arguments. (say \code{"method"} and \code{"jitter.scale"}
in \code{"resids"} function. Just keep them as default if not specified. Currently ignored.)}
}
\value{
An object of class \code{"PAsso"} is a list containing at least the following
components if \code{"partial"} is the association argument. It contains the partial
correlation matrix and multiple repeats if \code{rep_num} > 1. This object has "arguments"
attribute saved as c(association, method, resids.method), "responses" attribute, and
"adjustments" attribute.
The list contains:
\describe{
  \item{\code{corr}}{The estimated correlation matrix(average of \code{rep_MatCorr}) of partial association after
  adjusting confounders;}
  \item{\code{rep_corr}}{The replications of estimated correlation matrix;}
  \item{\code{rep_SRs}}{The replications of surrogate residuals if partial association is applied;}
  \item{\code{fitted.models}}{The list stores all fitted.models;}
  \item{\code{data}}{The data.frame of original dataset;}
  \item{\code{mods_n}}{The sample size of each fitted model;}
  \item{\code{cor_func}}{The correlation function after assign different method.}
}
}
\description{
This function is mainly designed for conducting the partial association analysis
bewteen two or more ordinal response variables after adjusting for a set of
covariates. It can also conduct marginal correlation analyses without adjusting
covariates. It generates an object that has correlation coefficients matrix,
and some attributes ("arguments" saves c(association, method, resids.method)).
The attribute "responses" contains the names of response variables.
}
\examples{
# Import nes2016_pre data in "PAsso"
data(nes2016_pre)
# Parial association:
PAsso_1 <- PAsso(responses = c("Prevote.num", "PID"),
                adjustments = c("income.num", "age", "edu.year"),
                data = nes2016_pre,
                association = c("partial"),
                models = c("probit", "probit"),
                method = c("kendall"))

PAsso_1

}
\references{
Liu, Dungang, Li, Shaobo, Yu, Yu, and Moustaki, Irini. Assessing partial association between
ordinal variables: quantification, visualization, and hypothesis testing, \emph{Journal of the
American Statistical Association}, Revision under review.

Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association}.
\url{http://www.tandfonline.com/doi/abs/10.1080/01621459.2017.1292915?journalCode=uasa20}
}
