% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_SR.def.R
\name{Pcor_SR.def}
\alias{Pcor_SR.def}
\title{Pcor_SR.def}
\usage{
Pcor_SR.def(var, confounder, data, links = c("probit", "probit"),
  asso_type = c("PartialAsso", "Marginal"), cor_method = c("kendall",
  "pearson", "wolfsigma"), surro_method = c("latent", "jitter", "Sign",
  "General", "Deviance"), rep_num = 100, ...)
}
\arguments{
\item{var}{a string vector that specifies response variables.}

\item{confounder}{a string vector that specifies confounders.}

\item{data}{the data.frame including responses and confounders.}

\item{links}{a string vector contains default link function for fitting models on
the input response variables.}

\item{asso_type}{a character string indicating whether the partial association coefficient
\code{"PartialAsso"} (default) or the marginal correlation \code{"Marginal"} is to be
computed. They can be abbreviated.}

\item{cor_method}{correlation coefficient method. Three choices \code{c("kendall", "pearson", "wolfsigma")},
and the default is \code{"kendall"}}

\item{surro_method}{Character string specifying which method to use to generate the
surrogate response values. Current options are \code{"latent"} and
\code{"jitter"}. Default is \code{"latent"}.
\describe{
  \item{\code{latent}}{surrogate approach;}
  \item{\code{jitter}}{jittering approach;}
  \item{\code{Sign}}{Sign-based residuals;}
  \item{\code{General}}{Generalized residuals (Franses and Paap 2001);}
  \item{\code{Deviance}}{Deviance residuals (-2*loglik).}
}}

\item{rep_num}{the repeat times for getting the average as estimator of partial
association pho. It is the \code{"nsim"} in \code{"resids()"} function}

\item{...}{Additional optional arguments. (say \code{"method"} and \code{"jitter.scale"}
in \code{"resids"} function. Just keep them as default if not specified. Currently ignored.)}
}
\value{
A list of class \code{c("PartialCor")} containing
the partial correlation matrix and multiple repeats if \code{rep_num} > 1.
This object has "arguments" attribute to save c(asso_type, cor_method, surro_method).
The list contains:
\describe{MatCor=MatCor, rep_MatCor=rep_MatCor, rep_SRs=rep_SRs, models
  \item{\code{MatCor}}{The estimated correlation matrix(average of \code{rep_MatCor}) of partial association controling confounders;}
  \item{\code{rep_MatCor}}{The replications of estimated correlation matrix;}
  \item{\code{rep_SRs}}{The replications of surrogate residuals if partial association is applied;}
  \item{\code{models}}{The list stores all input models.}
}
}
\description{
Pcor_SR.def
}
\examples{
# import data
data(nes96)
Pcortest_1 <- Pcor_SR.def(var = c("vote.num", "PID"),
confounder = c("income.num", "age", "edu.year"),
asso_type = "PartialAsso", cor_method = "kendall",
rep_num = 30, data = nes96, links=c("probit", "probit"))

}
