% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_SR.R
\name{Pcor_SR}
\alias{Pcor_SR}
\title{Pcor_SR
function to obtain \eqn{hat{phi}}}
\usage{
Pcor_SR(y1, y2, x, model1, model2, alpha1 = NULL, beta1 = NULL,
  alpha2 = NULL, beta2 = NULL, link1 = "probit", link2 = "probit",
  n.avg = 100, method = "kendall")
}
\arguments{
\item{y1}{ordinal response 1}

\item{y2}{ordinal response 2}

\item{x}{covariates}

\item{model1}{The model 1 between ordinal response 1 and covariates}

\item{model2}{The model 2 between ordinal response 1 and covariates}

\item{alpha1}{Alpha of model 1}

\item{beta1}{Betas of model 1}

\item{alpha2}{Alpha of model 2}

\item{beta2}{Betas of model 2}

\item{link1}{link function of ordinal response 1}

\item{link2}{link function of ordinal response 2}

\item{n.avg}{The number of simulation for generate residuals}

\item{method}{The method to calculate correlation}
}
\value{
A list contains ...
\itemize{
 \item{"parameter 1"}{Stuff}
 \item{"parameter 2"}{Stuff}
}
}
\description{
Pcor_SR
function to obtain \eqn{hat{phi}}
}
