% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_SR.test.R
\name{Pcor_SR.test}
\alias{Pcor_SR.test}
\title{Pcor_SR.test}
\usage{
Pcor_SR.test(object, boot_SE = 100, H0 = 0, parallel = FALSE)
}
\arguments{
\item{object}{an object of "PartialCor" class, which is generated by "Pcor_SR.def" function.}

\item{boot_SE}{bootstrap replications. It may be very slow for partial association
so a double-check question should be answered before move on.}

\item{H0}{null hypothesis of partial correlation coefficient.}

\item{parallel}{logical argument whether conduct parallel for bootstrapping partial association.}
}
\description{
Pcor_SR.test
}
