% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcor_SR_boot.R
\name{Pcor_SR_boot}
\alias{Pcor_SR_boot}
\title{Pcor_SR_boot function to perform bootstrap inference}
\usage{
Pcor_SR_boot(y1, y2, x, B = 2000, link1 = "probit", link2 = "probit",
  n.avg = 100, method = "kendall", H0 = 0)
}
\arguments{
\item{y1}{ordinal response 1}

\item{y2}{ordinal response 2}

\item{x}{covariates}

\item{B}{bootstrap times}

\item{link1}{link function of ordinal response 1}

\item{link2}{link function of ordinal response 2}

\item{n.avg}{The number of simulation for generate residuals}

\item{method}{The method to calculate correlation}

\item{H0}{The mu when conduct hypothesis testing}
}
\value{
A list contains ...
}
\description{
Pcor_SR_boot function to perform bootstrap inference
}
