% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\alias{plot3D.default}
\alias{plot3D.PAsso}
\title{3-D P-P plot for the inspection of the partial association analysis}
\usage{
plot3D(object)

\method{plot3D}{default}(object, ...)

\method{plot3D}{PAsso}(object, ...)
}
\arguments{
\item{object}{The input object should be a "PAsso" class that is generated
by "Passo" or "test".}
}
\value{
A list of \code{"plotly"} objects with name "plot_1", "plot_2" etc.
}
\description{
A list of 3-D P-P plots for the inspection of the partial
association analysis. Each plot is the 3-D P-P plot from an empirical
copula trained from the surrogate residuals of a pair of responses.
}
\details{
All the plots are based on surrogate residuals generated from \code{"resides"}
function in \code{\link{sure}}. Graphics are designed based on
\code{\link{PAsso}} and \code{"plotly"}.
}
\examples{
data("nes2016")
PAsso_3v <- PAsso(responses = c("vote.num", "PID", "selfLR"),
                 adjustments = c("income.num", "age", "edu.year"),
                 data = nes2016)

plot_all <- plot3D(PAsso_3v)
plot_all$plot_1

}
