% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D_one}
\alias{plot3D_one}
\title{plot3D_one}
\usage{
plot3D_one(
  plot_list,
  rep_SRs,
  m,
  n,
  plot_titles,
  type = c("surface3D", "contour")
)
}
\arguments{
\item{plot_list}{The list to save plots. Each 3D plot is between one pair of covariates.}

\item{rep_SRs}{The surrogate responses array saved in the PAsso object.}

\item{m}{The index of first covariate. Since the names of covariates are empty, need to input index number.}

\item{n}{The index of second covariate.}

\item{plot_titles}{The title of the plot.}

\item{type}{A character string specifying the trace type (e.g. "surface3D", "contour"). "contour" creates a
2D contour plot between u and v.}
}
\description{
plot3D_one
}
\keyword{(3D}
\keyword{This}
\keyword{an}
\keyword{contour}
\keyword{draw}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{one}
\keyword{or}
\keyword{plot).}
\keyword{plotly}
\keyword{surface}
\keyword{to}
