% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals,vglm-method}
\alias{residuals,vglm-method}
\title{ref the vglm' adjacent categories regression model to using \code{"residualsAcat"} function.}
\usage{
\S4method{residuals}{vglm}(object, ...)
}
\description{
ref the vglm' adjacent categories regression model to using \code{"residualsAcat"} function.
}
