% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residualsVGAM}
\alias{residualsVGAM}
\title{Extract Residuals from models of VGAM}
\usage{
residualsVGAM(
  object,
  type = c("surrogate", "sign", "general", "deviance"),
  jitter = c("latent", "uniform"),
  jitter.uniform.scale = c("probability", "response"),
  nsim = 1L,
  ...
)
}
\description{
A internal function to simulate surrogate residuals for models fitted by
\code{\link[VGAM]{vglm}} and \code{\link[VGAM]{vglm}}. Now, this one support \code{"vglm"} and
\code{"vgam"}, and adjacent categories regression model by \code{"vglm"}. This one may
need to update to support more models from VGAM.
}
\keyword{internal}
