% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuRes.R
\name{SR}
\alias{SR}
\title{Surrogate Residual Generating}
\usage{
SR(y, X, model, alpha = NULL, beta = NULL, ndraw = 1,
  dist = "norm")
}
\arguments{
\item{y}{ordinal response variable}

\item{X}{covariates that need to be adjusted}

\item{model}{The cumulative link model formula}

\item{alpha}{The cumulative intercepts}

\item{beta}{The coefficients in front of covariates}

\item{ndraw}{The number of simulation for generate residuals}

\item{dist}{The argument to specify distribution}
}
\value{
A vector contains surrogate residuals for the binary response y
}
\description{
This function simulates surrogate residuals for ordinal variables by adjusting
a set of covariates X using cumulative link models. Various links are considered.
This version incorporates binary response from glm model.
}
