% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAsso.R
\name{summary}
\alias{summary}
\alias{summary.PAsso}
\title{Summary of partial association analysis}
\usage{
\method{summary}{PAsso}(object, digits = max(3L, getOption("digits") - 2L), ...)
}
\description{
This function summarizes the partial association analysis by
providing partial association matrix, marginal association matrix, and a
matrix summarizing the model's coefficients. The partial correlation coefficient
matrix displays the partial association between each pair of responses
after adjusting the covariates. While the marginal coefficient matrix displays association
before the adjustment.
}
\examples{
# See PAsso for the example.

}
