% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Hypothesis testing of the partial association coefficients}
\usage{
test(object, boot_SE = 300, H0 = 0, parallel = FALSE)
}
\arguments{
\item{object}{An object of "PAsso" class, which is generated by "PAsso" function.}

\item{boot_SE}{The number of bootstrap replications. It may be very slow for partial
association so a double-check question will ask before move on.}

\item{H0}{null hypothesis of partial correlation coefficient.}

\item{parallel}{logical argument whether conduct parallel for bootstrapping partial association.}
}
\description{
This function use bootstrapping to conduct hypothesis testing
for the partial association coefficients. It can directly apply onto the "PAsso"
class of object generated by "PAsso".
}
\examples{
# Import nes2016_pre data in "parasol"
data(nes2016_pre)
# Parial association:
PAsso_1 <- PAsso(responses = c("Prevote.num", "PID"),
                adjustments = c("income.num", "age", "edu.year"),
                data = nes2016_pre,
                association = c("partial"))

summary(PAsso_1, digits=4)

PAsso_1_test <- test(object = PAsso_1, boot_SE=20, H0=0, parallel=FALSE))

}
