% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{The Correlation and partial correlation coefficients between
ordinal responses after adjusting for a set of covariates.}
\usage{
corr(responses, adjustments, data, association = c("partial",
  "marginal"), models = c("probit", "probit"), method = c("kendall",
  "pearson", "wolfsigma"), resids.method = c("latent", "jitter", "Sign",
  "General", "Deviance"), fitted.models = NULL, rep_num = 100, ...)
}
\arguments{
\item{responses}{A string vector that specifies response variables. It requires to be equal
or greater than two variables in the data frame.}

\item{adjustments}{A string vector specifies covariates/confounders that need to
be adjusted.}

\item{data}{the data.frame including responses and adjustments.}

\item{association}{A character string indicating whether the partial association
coefficient \code{"partial"} (default) or the marginal correlation \code{"marginal"}
is to be computed. They can be abbreviated.}

\item{models}{A string vector contains default link function of fitting models with
respect to each response variable.}

\item{method}{A string argument to specify correlation coefficient method.
Three choices \code{c("kendall", "pearson", "wolfsigma")}. The default is
\code{"kendall"}}

\item{resids.method}{A character string specifying which method to use to generate the
residuals. Current options are \code{"latent"} and
\code{"jitter"}. Default is \code{"latent"}.
\describe{
  \item{\code{latent}}{surrogate approach;}
  \item{\code{jitter}}{jittering approach;}
  \item{\code{Sign}}{Sign-based residuals;}
  \item{\code{General}}{Generalized residuals (Franses and Paap 2001);}
  \item{\code{Deviance}}{Deviance residuals (-2*loglik).}
}}

\item{fitted.models}{A list contains all the models (S3 objects) you want to
assess for the partial association between ordinal responses after adjusting
for a set of covariates covariates. All of these models should be applied to the
same dataset, having same covariates, same sample size etc. The models in this
list can be an object of class \code{\link[ordinal]{clm}},
\code{\link[stats]{glm}}, \code{\link[rms]{lrm}}, \code{\link[rms]{orm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{rep_num}{A number to specify repeat times of simulation of surrogate residuls
such that the partial correlation coefficients are calculated repeatly. The final
correlation coefficents the average of all partial correlation coefficients.
It is the \code{"nsim"} argument in \code{"resids()"} function.}

\item{...}{Additional optional arguments. (say \code{"method"} and \code{"jitter.scale"}
in \code{"resids"} function. Just keep them as default if not specified. Currently ignored.)}
}
\value{
An object of class \code{"PAsso"} is a list containing at least the following
components if \code{"partial"} is the association argument. It contains the partial
correlation matrix and multiple repeats if \code{rep_num} > 1. This object has "arguments"
attributes saved as c(association, method, resids.method).
The list contains:
\describe{
  \item{\code{corr}}{The estimated correlation matrix(average of \code{rep_MatCorr}) of partial association after
  adjusting confounders;}
  \item{\code{rep_corr}}{The replications of estimated correlation matrix;}
  \item{\code{rep_SRs}}{The replications of surrogate residuals if partial association is applied;}
  \item{\code{fitted.models}}{The list stores all fitted.models;}
  \item{\code{data}}{The data.frame of original dataset;}
  \item{\code{mods_n}}{The sample size of each fitted model;}
  \item{\code{cor_func}}{The correlation function after assign different method.}
}
}
\description{
This function is mainly designed for conducting the partial association analysis
bewteen two or more ordinal response variables after adjusting for a set of
covariates. It can also conduct marginal correlation analyses without adjusting
covariates. It generates an object that has correlation coefficients matrix,
and some attributes ("arguments" saves c(association, method, resids.method)).
The attribute "responses" contains the names of response variables.
}
\examples{
#
# Adjacent Categories Regression Model Example to compare different residuals
#

data("df_ParA")
summary(df_ParA$data)
fit_clm1 <- VGAM::vglm(Y1 ~ X, family =
                      VGAM::cumulative(link = "logit",reverse=TRUE,parallel = TRUE),
                      data = df_ParA$data)
fit_clm2 <- VGAM::vglm(Y2 ~ X, family =
                       VGAM::cumulative(link = "logit",reverse=TRUE,parallel = TRUE),
                       data = df_ParA$data)
SR1 <- resids(fit_clm1, method = "latent",boot_id = NULL)
SR2 <- resids(fit_clm2, method = "latent", boot_id = NULL)

## obtain SBC residuals (Li and Shepherd 2012 JASA/Biometrika)
PR1 <- resids(fit_clm1, method = "Sign", boot_id = NULL)
PR2 <- resids(fit_clm2, method = "Sign", boot_id = NULL)

## obtain generalized residuals (Franses and Paap 2001 book)
GR1 <- resids(fit_clm1, method = "General", boot_id = NULL)
GR2 <- resids(fit_clm2, method = "General", boot_id = NULL)

## obtain deviance residuals
DR1 <- resids(fit_clm1, method = "Deviance", boot_id = NULL)
DR2 <- resids(fit_clm2, method = "Deviance", boot_id = NULL)

## visualize residual vs. residual
par(mfrow=c(2,2))
par(mar=c(4, 4.8, 2.5, 1.5))

plot(PR1, PR2, pch=".", main = "Sign-based Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")))
plot(GR1, GR2, pch=".", main = "Generalized Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")), xlim = c(-4,4), ylim=c(-4,4))
plot(DR1, DR2, pch='.', main = "Deviance Residuals",
     xlab = expression(paste(R[1]^"ALT")),
     ylab = expression(paste(R[2]^"ALT")))
plot(SR1, SR2, pch=".", main = "Surrogate Residuals", xaxt="n", yaxt="n",
     xlab = expression(R[1]), ylab = expression(R[2]),
     xlim = c(-1/2,1/2), ylim=c(-1/2,1/2))
axis(1, at=seq(-0.5, 0.5, 0.25), labels = seq(-0.5, 0.5, 0.25))
axis(2, at=seq(-0.5, 0.5, 0.25), labels = seq(-0.5, 0.5, 0.25))

# Import nes96 data in "parasol"
data(nes96)
# Parial association:
PAsso_1 <- corr(responses = c("vote.num", "PID"),
                adjustments = c("income.num", "age", "edu.year"),
                data = nes96,
                association = c("partial"),
                models = c("probit", "probit"),
                method = c("kendall"),
                resids.method = "latent",
                fitted.models = NULL, rep_num = 100)

# Marginal association:
MAsso_1 <- corr(responses = c("vote.num", "PID"),
                adjustments = c("income.num", "age", "edu.year"),
                data = nes96,
                association = c("marginal"))

# Compare marginal correlation with partial correlation.
PAsso_1
MAsso_1


}
\references{
Dungang Liu, Shaobo Li and Yan Yu. Assessing Partial Association Between Ordinal
Variables: A General Framework

Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted). URL
http://www.tandfonline.com/doi/abs/10.1080/01621459.2017.1292915?journalCode=uasa20
}
