% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.test.R
\name{corr.test}
\alias{corr.test}
\title{Hypothesis testing of the partial correlation coefficients by bootstrapping the approach in "corr".}
\usage{
corr.test(object, boot_SE = 300, H0 = 0, parallel = FALSE)
}
\arguments{
\item{object}{An object of "PAsso"/"MAsso" class, which is generated by "corr.def" function.}

\item{boot_SE}{The number of bootstrap replications. It may be very slow for partial
association so a double-check question will ask before move on.}

\item{H0}{null hypothesis of partial correlation coefficient.}

\item{parallel}{logical argument whether conduct parallel for bootstrapping partial association.}
}
\description{
This function use bootstrap to conduct hypothesis testing for the partial correlation
coefficients. It can directly apply onto the object generated by "corr".
}
