% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_AdjCat}
\alias{df_AdjCat}
\title{Ordinal Response with Adjacent Categories Probabilities}
\format{
A data frame with 10000 rows and 11 variables.
\itemize{
  \item \code{X} A continuous predictor variable.
  \item \code{Y1} The response variable; an ordered factor.
  \item \code{Y2} The response variable; an ordered factor.
}
}
\usage{
data(df_AdjCat)
}
\description{
Data simulated from an adjacent categories regression model
with an ordered (preferably) factor response. beta1 = 1, beta2 = -1,
alpha1 = (-Inf, -3, -2, 0, 2, 3, Inf), alpha2 = (-Inf, -2, 0, 2, Inf)
}
\examples{
head(df_AdjCat)
}
\references{
Liu, Dungang, Li, Shaobo, Yu, Yu, and Moustaki, Irini. Assessing partial association between
ordinal variables: quantification, visualization, and hypothesis testing, \emph{Journal of the
American Statistical Association}, Revision under review.
}
\keyword{datasets}
