% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{p_adj_cate}
\alias{p_adj_cate}
\title{p_adj_cate}
\usage{
p_adj_cate(Z)
}
\arguments{
\item{Z}{A numerical vector that inputs the latent variable for generating probabilities of adjacent
categories regression model.}
}
\value{
A matrix (n by level of respones plus 1) of probabilities of the adjacent categories model.
}
\description{
p_adj_cate
}
\keyword{internal}
