% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.PAsso}
\title{A matrix of partial regression plots between responses after adjustments}
\usage{
plot(x, y, ...)

\method{plot}{PAsso}(x, color = "blue", ...)
}
\arguments{
\item{x}{A PAsso object for drawing partial regression plot.}

\item{...}{}

\item{color}{The color of points.}
}
\value{
A \code{"GGally"} object.
}
\description{
A plot matrix to display the results of partial association analyses. Upper-triangle contains
scatter-plot matrix between each pair of response variables. Lower-triangle contains the
partial correlation coefficients adjusted by covariates.
}
\details{
A pairwise plot matrix reveals the partial association between ordinal variables.
All the plots are based on surrogate residuals generated from \code{"resides"} function.
Graphics are designed based on \code{\link{ggplot2}} and \code{"GGally"}.
}
\examples{
data(nes2016_pre)

summary(nes2016_pre)

PAsso_1 <- PAsso(responses = c("Prevote.num", "PID"),
                 adjustments = c("income.num", "age", "edu.year"),
                 data = nes2016_pre)

plot(x = PAsso_1)

}
