% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\alias{plot3D.default}
\alias{plot3D.PAsso}
\title{plot3D: The 3-D P-P plot for partial association analysis}
\usage{
plot3D(object)

\method{plot3D}{default}(x, ...)

\method{plot3D}{PAsso}(object, ...)
}
\arguments{
\item{object}{The input object should be a "PAsso" class that is generated by "Passo" or "test".}
}
\value{
A list of \code{"plotly"} object with name "plot_1", "plot_2" etc.

\code{NULL}

\code{NULL}
}
\description{
A 3-D P-P plot to inspect the partial association analyses
}
\details{
A list of plots to inspect the results of partial association analyses.
Each plot is the 3-D P-P plot from an empirical copula trained from the
surrogate residuals of a pair of responses.

All the plots are based on surrogate residuals generated from \code{"resides"}
function in \code{\link{sure}}. Graphics are designed based on
\code{\link{PAsso}} and \code{"plotly"}.
}
\examples{
data("nes2016_pre")
PAsso_2 <- PAsso(responses = c("vote.num", "PID", "selfLR"),
                 adjustments = c("income.num", "age", "edu.year"),
                 data = nes2016_pre)

plot_all <- plot3D(PAsso_2)
plot_all$plot_1

}
