% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAsso.R, R/test.R
\name{print}
\alias{print}
\alias{print.PAsso}
\alias{print.PAsso.test}
\title{Print partial association matrix}
\usage{
\method{print}{PAsso}(x, digits = max(2, getOption("digits") - 2), ...)

\method{print}{PAsso.test}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A PAsso object for printing out results.}

\item{digits}{A default number to specify decimal digit values.}

\item{...}{Additional optional arguments.}
}
\value{
Print partial association matrix of a PAsso object
}
\description{
Print partial association matrix
}
\examples{
# See PAsso for the example.

}
