% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAsso.R
\name{summary.PAsso}
\alias{summary.PAsso}
\title{Summary of partial association analysis}
\usage{
\method{summary}{PAsso}(object, digits = max(3L, getOption("digits") - 2L), ...)
}
\description{
This function summarizes the partial association analysis by
providing partial association matrix, marginal association matrix, and a
matrix of models' estimation.
}
